+++ tests/mir_codegen/try_else/main.c
@@ -7,9 +7,7 @@ int main(void) {
     struct Pair p = maybe_fail(1);
     if (p.err) {
         const char* msg = error_to_cstr(p.err);
-        if (msg) {
-            fprintf(stderr, "%s\n", msg);
-        }
+        if (msg) fprintf(stderr, "%s\n", msg);
         error_free(p.err);
         return 1;
     }